+++ Module.php
@@ -31,6 +31,7 @@ namespace Search;
 
 use Laminas\ModuleManager\ModuleManager;
 use Laminas\EventManager\SharedEventManagerInterface;
+use Laminas\EventManager\Event;
 use Laminas\Mvc\MvcEvent;
 use Laminas\ServiceManager\ServiceLocatorInterface;
 use Omeka\Module\AbstractModule;
@@ -38,6 +39,8 @@ use Composer\Semver\Comparator;
 
 class Module extends AbstractModule
 {
+    protected $resourcesBeingUpdatedInBatch = [];
+
     public function getConfig()
     {
         return include __DIR__ . '/config/module.config.php';
@@ -169,50 +172,176 @@ class Module extends AbstractModule
 
     public function attachListeners(SharedEventManagerInterface $sharedEventManager)
     {
-        $sharedEventManager->attach(
-            'Omeka\Api\Adapter\ItemAdapter',
-            'api.create.post',
-            [$this, 'updateSearchIndex']
-        );
-        $sharedEventManager->attach(
-            'Omeka\Api\Adapter\ItemAdapter',
-            'api.update.post',
-            [$this, 'updateSearchIndex']
-        );
-        $sharedEventManager->attach(
-            'Omeka\Api\Adapter\ItemAdapter',
-            'api.delete.post',
-            [$this, 'updateSearchIndex']
-        );
-
-        $sharedEventManager->attach(
-            'Omeka\Api\Adapter\ItemSetAdapter',
-            'api.create.post',
-            [$this, 'updateSearchIndex']
-        );
-        $sharedEventManager->attach(
-            'Omeka\Api\Adapter\ItemSetAdapter',
-            'api.update.post',
-            [$this, 'updateSearchIndex']
-        );
-        $sharedEventManager->attach(
-            'Omeka\Api\Adapter\ItemSetAdapter',
-            'api.delete.post',
-            [$this, 'updateSearchIndex']
-        );
+        $identifiers = ['Omeka\Api\Adapter\ItemAdapter', 'Omeka\Api\Adapter\ItemSetAdapter'];
+        foreach ($identifiers as $identifier) {
+            $sharedEventManager->attach(
+                $identifier,
+                'api.batch_update.pre',
+                [$this, 'onResourceBatchUpdatePre']
+            );
+            $sharedEventManager->attach(
+                $identifier,
+                'api.batch_update.post',
+                [$this, 'onResourceBatchUpdatePost']
+            );
+            $sharedEventManager->attach(
+                $identifier,
+                'api.update.post',
+                [$this, 'onResourceUpdatePost']
+            );
+            $sharedEventManager->attach(
+                $identifier,
+                'api.batch_create.pre',
+                [$this, 'onResourceBatchCreatePre']
+            );
+            $sharedEventManager->attach(
+                $identifier,
+                'api.batch_create.post',
+                [$this, 'onResourceBatchCreatePost']
+            );
+            $sharedEventManager->attach(
+                $identifier,
+                'api.create.post',
+                [$this, 'onResourceCreatePost']
+            );
+            $sharedEventManager->attach(
+                $identifier,
+                'api.delete.post',
+                [$this, 'onResourceDeletePost']
+            );
+        }
     }
 
-    public function updateSearchIndex($event)
+    public function onResourceBatchUpdatePre(Event $event)
+    {
+        $request = $event->getParam('request');
+        $ids = $request->getIds();
+        foreach ($ids as $id) {
+            $key = sprintf('%s:%s', $request->getResource(), $id);
+            $this->resourcesBeingUpdatedInBatch[$key] = true;
+        }
+    }
+
+    public function onResourceBatchUpdatePost(Event $event)
+    {
+        $response = $event->getParam('response');
+        $resources = $response->getContent();
+        $this->indexResources($resources);
+
+        foreach ($resources as $resource) {
+            $key = sprintf('%s:%s', $resource->getResourceName(), $resource->getId());
+            unset($this->resourcesBeingUpdatedInBatch[$key]);
+        }
+    }
+
+    public function onResourceUpdatePost(Event $event)
+    {
+        $request = $event->getParam('request');
+        $key = sprintf('%s:%s', $request->getResource(), $request->getId());
+        if (array_key_exists($key, $this->resourcesBeingUpdatedInBatch)) {
+            // Do nothing. These resources will be indexed in
+            // api.batch_update.post listener
+            return;
+        }
+
+        $response = $event->getParam('response');
+        $resource = $response->getContent();
+        $this->indexResources([$resource]);
+    }
+
+    public function onResourceBatchCreatePre(Event $event)
+    {
+        $request = $event->getParam('request');
+        $content = $request->getContent();
+        foreach ($content as &$c) {
+            // This is used in api.create.post listener to avoid indexing the
+            // same resources twice
+            $c['o-module-search:batch-create'] = true;
+        }
+        $request->setContent($content);
+    }
+
+    public function onResourceBatchCreatePost(Event $event)
+    {
+        $response = $event->getParam('response');
+        $resources = $response->getContent();
+        $this->indexResources($resources);
+    }
+
+    public function onResourceCreatePost(Event $event)
+    {
+        $request = $event->getParam('request');
+        $content = $request->getContent();
+        if (array_key_exists('o-module-search:batch-create', $content)) {
+            // Do nothing. These resources will be indexed in
+            // api.batch_create.post listener
+            return;
+        }
+
+        $response = $event->getParam('response');
+        $resource = $response->getContent();
+        $this->indexResources([$resource]);
+    }
+
+    public function onResourceDeletePost(Event $event)
     {
         $serviceLocator = $this->getServiceLocator();
-
+        $api = $serviceLocator->get('Omeka\ApiManager');
+        $logger = $serviceLocator->get('Omeka\Logger');
         $request = $event->getParam('request');
-        $response = $event->getParam('response');
-        $resourceId = $request->getId() ? $request->getId() : $response->getContent()->getId();
 
+        $searchIndexes = $api->search('search_indexes')->getContent();
+        foreach ($searchIndexes as $searchIndex) {
+            $searchIndexSettings = $searchIndex->settings();
+            if (!in_array($request->getResource(), $searchIndexSettings['resources'])) {
+                continue;
+            }
+
+            try {
+                $indexer = $searchIndex->indexer();
+                $indexer->deleteResource($request->getResource(), $request->getId());
+            } catch (\Exception $e) {
+                $logger->err(sprintf('Search: failed to delete resource: %s', $e));
+            }
+        }
+    }
+
+    protected function indexResources(array $resources)
+    {
+        if (empty($resources)) {
+            return;
+        }
+
+        $serviceLocator = $this->getServiceLocator();
+        $logger = $serviceLocator->get('Omeka\Logger');
         $jobDispatcher = $serviceLocator->get(\Omeka\Job\Dispatcher::class);
-        $jobClass = $request->getOperation() === 'delete' ? 'Search\Job\DeleteIndexSingle' : 'Search\Job\IndexSingle';
-        $jobDispatcher->dispatch($jobClass, ['id' => $resourceId, 'type' => $request->getResource()]);
+
+        // If we are in a background job or a script executed from the command
+        // line, we do not have a time limit so we can index resources
+        // immediately.
+        // Otherwise, index resources in a background job in order to not slow
+        // down the request
+        if (PHP_SAPI === 'cli') {
+            $api = $serviceLocator->get('Omeka\ApiManager');
+            $searchIndexes = $api->search('search_indexes')->getContent();
+            foreach ($searchIndexes as $searchIndex) {
+                $searchIndexSettings = $searchIndex->settings();
+                $filteredResources = array_filter($resources, fn($resource) => in_array($resource->getResourceName(), $searchIndexSettings['resources']));
+                if (empty($filteredResources)) {
+                    continue;
+                }
+
+                try {
+                    $indexer = $searchIndex->indexer();
+                    $indexer->indexResources($filteredResources);
+                } catch (\Exception $e) {
+                    $logger->err(sprintf('Search: failed to index resources: %s', $e));
+                }
+            }
+        } else {
+            $ids = array_map(fn($resource) => $resource->getId(), $resources);
+            $jobDispatcher->dispatch(Job\UpdateIndex::class, ['ids' => $ids]);
+        }
     }
 
     protected function addRoutes()